+++ src/Middleware/DispatcherMiddleware.php
@@ -6,10 +6,12 @@ namespace WoohooLabs\Harmony\Middleware;
 use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Server\MiddlewareInterface;
+use Psr\Http\Server\RequestHandlerInterface;
 use WoohooLabs\Harmony\Container\BasicContainer;
 use WoohooLabs\Harmony\Exception\DispatcherException;
 
-class DispatcherMiddleware
+class DispatcherMiddleware implements MiddlewareInterface
 {
     /**
      * @var ContainerInterface
@@ -23,26 +25,25 @@ class DispatcherMiddleware
 
     public function __construct(ContainerInterface $container = null, string $actionAttributeName = "__action")
     {
-        $this->container = $container === null ? new BasicContainer() : $container;
+        $this->container = $container ?? new BasicContainer();
         $this->actionAttributeName = $actionAttributeName;
     }
 
     /**
      * @throws DispatcherException
      */
-    public function __invoke(
-        ServerRequestInterface $request,
-        ResponseInterface $response,
-        callable $next
-    ): ResponseInterface {
+    public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
+    {
         $action = $request->getAttribute($this->actionAttributeName);
 
         if ($action === null) {
             throw new DispatcherException(
-                "Please provide the '{$this->actionAttributeName}' attribute as a callable for the request object in " .
+                "Please set the '{$this->actionAttributeName}' attribute as a callable in the request object in " .
                 "order to be able dispatch it!"
             );
         }
+
+        $response = $handler->handle($request);
 
         if (is_array($action) && is_string($action[0]) && is_string($action[1])) {
             $object = $this->container->get($action[0]);
@@ -54,7 +55,7 @@ class DispatcherMiddleware
             $response = $action($request, $response);
         }
 
-        return $next($request, $response);
+        return $response;
     }
 
     public function getContainer(): ContainerInterface