+++ tests/Condition/ExactPathConditionTest.php
@@ -3,8 +3,8 @@ declare(strict_types=1);
 
 namespace WoohooLabs\Harmony\Tests\Container;
 
+use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use PHPUnit_Framework_MockObject_MockObject;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\UriInterface;
 use WoohooLabs\Harmony\Condition\ExactPathCondition;
@@ -17,11 +17,11 @@ class ExactPathConditionTest extends TestCase
      */
     public function evaluateToTrue()
     {
-        /** @var UriInterface|PHPUnit_Framework_MockObject_MockObject $uri */
+        /** @var UriInterface|MockObject $uri */
         $uri = $this->getMockBuilder(UriInterface::class)->getMock();
         $uri->method("getPath")->willReturn("/api");
 
-        /** @var ServerRequestInterface|PHPUnit_Framework_MockObject_MockObject $request */
+        /** @var ServerRequestInterface|MockObject $request */
         $request = $this->getMockBuilder(ServerRequestInterface::class)->getMock();
         $request->method("getUri")->willReturn($uri);
 
@@ -35,11 +35,11 @@ class ExactPathConditionTest extends TestCase
      */
     public function evaluateToFalse()
     {
-        /** @var UriInterface|PHPUnit_Framework_MockObject_MockObject $uri */
+        /** @var UriInterface|MockObject $uri */
         $uri = $this->getMockBuilder(UriInterface::class)->getMock();
         $uri->method("getPath")->willReturn("/api");
 
-        /** @var ServerRequestInterface|PHPUnit_Framework_MockObject_MockObject $request */
+        /** @var ServerRequestInterface|MockObject $request */
         $request = $this->getMockBuilder(ServerRequestInterface::class)->getMock();
         $request->method("getUri")->willReturn($uri);