+++ tests/Condition/PathPrefixConditionTest.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace WoohooLabs\Harmony\Tests\Container;
 
+use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use PHPUnit_Framework_MockObject_MockObject;
 use Psr\Http\Message\ServerRequestInterface;
@@ -17,11 +18,11 @@ class PathPrefixConditionTest extends TestCase
      */
     public function evaluateExactPathToTrue()
     {
-        /** @var UriInterface|PHPUnit_Framework_MockObject_MockObject $uri */
+        /** @var UriInterface|MockObject $uri */
         $uri = $this->getMockBuilder(UriInterface::class)->getMock();
         $uri->method("getPath")->willReturn("/api");
 
-        /** @var ServerRequestInterface|PHPUnit_Framework_MockObject_MockObject $request */
+        /** @var ServerRequestInterface|MockObject $request */
         $request = $this->getMockBuilder(ServerRequestInterface::class)->getMock();
         $request->method("getUri")->willReturn($uri);
 
@@ -35,7 +36,7 @@ class PathPrefixConditionTest extends TestCase
      */
     public function evaluateSubPathToTrue()
     {
-        /** @var UriInterface|PHPUnit_Framework_MockObject_MockObject $uri */
+        /** @var UriInterface|MockObject $uri */
         $uri = $this->getMockBuilder(UriInterface::class)->getMock();
         $uri->method("getPath")->willReturn("/api/users");