+++ tests/Utils/Middleware/ExceptionMiddleware.php
@@ -5,9 +5,11 @@ namespace WoohooLabs\Harmony\Tests\Utils\Middleware;
 
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Server\MiddlewareInterface;
+use Psr\Http\Server\RequestHandlerInterface;
 use WoohooLabs\Harmony\Tests\Utils\Exception\TestException;
 
-class ExceptionMiddleware
+class ExceptionMiddleware implements MiddlewareInterface
 {
     protected $text;
 
@@ -24,11 +26,8 @@ class ExceptionMiddleware
     /**
      * @throws TestException
      */
-    public function __invoke(
-        ServerRequestInterface $request,
-        ResponseInterface $response,
-        callable $next
-    ): ResponseInterface {
+    public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
+    {
         throw new TestException($this->text);
     }
 }