+++ tests/Utils/Middleware/HeaderMiddleware.php
@@ -5,8 +5,10 @@ namespace WoohooLabs\Harmony\Tests\Utils\Middleware;
 
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Server\MiddlewareInterface;
+use Psr\Http\Server\RequestHandlerInterface;
 
-class HeaderMiddleware
+class HeaderMiddleware implements MiddlewareInterface
 {
     /**
      * @var string
@@ -27,11 +29,10 @@ class HeaderMiddleware
         $this->value = $value;
     }
 
-    public function __invoke(
-        ServerRequestInterface $request,
-        ResponseInterface $response,
-        callable $next
-    ): ResponseInterface {
-        return $next(null, $response->withHeader($this->name, $this->value));
+    public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
+    {
+        $response = $handler->handle($request);
+
+        return $response->withHeader($this->name, $this->value);
     }
 }